@@ -3,59 +3,48 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
-    "test:watch": "vue-cli-service test:unit --watch",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src"
   },
   "dependencies": {
-    "@vue/compat": "^3.1.0",
-    "async": "^3.2.3",
-    "bootstrap": "4.6.1",
-    "bootstrap-vue": "^2.22.0",
-    "core-js": "^3.8.3",
+    "@vue/compat": "^3.2.45",
+    "async": "^3.2.4",
+    "bootstrap": "4.6.2",
+    "bootstrap-vue": "^2.23.1",
     "lodash": "^4.17.21",
-    "luxon": "^2.4.0",
-    "ol": "^6.14.1",
+    "luxon": "^2.5.0",
+    "ol": "^6.15.1",
     "popper.js": "^1.16.1",
-    "portal-vue": "^2.1.7",
+    "portal-vue": "^3.0.0",
     "sprintf-js": "^1.1.2",
-    "vue": "^3.1.0",
-    "vue-router": "3",
-    "vuex": "3"
+    "vue": "^3.2.45",
+    "vue-router": "^4.1.6",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/eslint-parser": "^7.12.16",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/user-event": "^14.2.0",
-    "@testing-library/vue": "5",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "~5.0.0",
-    "@vue/cli-plugin-router": "^5.0.4",
-    "@vue/cli-plugin-unit-jest": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "@vue/eslint-config-standard": "^7.0.0",
-    "@vue/test-utils": "^1.1.3",
-    "@vue/vue2-jest": "27",
-    "babel-core": "bridge",
-    "babel-jest": "^27.0.6",
-    "eslint": "^7.32.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/user-event": "^14.4.3",
+    "@testing-library/vue": "^6.6.1",
+    "@vitejs/plugin-vue": "^4.0.0",
+    "@vue/eslint-config-standard": "^8.0.1",
+    "@vue/test-utils": "^2.2.6",
+    "@vue/vue3-jest": "^29.2.2",
+    "eslint": "8.30.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.2.0",
-    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-n": "^15.6.0",
+    "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-standard": "^5.0.0",
-    "eslint-plugin-vue": "^8.0.3",
-    "jest": "^27.0.5",
+    "eslint-plugin-vue": "^9.8.0",
+    "jest": "^29.3.1",
     "jest-fetch-mock": "^3.0.3",
     "mutationobserver-shim": "^0.3.7",
-    "portal-vue": "^2.1.7",
-    "sass": "^1.51.0",
-    "sass-loader": "^12.6.0",
+    "portal-vue": "^3.0.0",
+    "sass": "^1.57.1",
+    "vite": "^4.0.3",
     "vue-cli": "^2.9.6",
-    "vue-cli-plugin-bootstrap-vue": "^0.8.2",
-    "vue-template-compiler": "^2.6.14"
+    "vue-cli-plugin-bootstrap-vue": "^0.8.2"
   },
   "packageManager": "yarn@3.0.1"
 }
