@@ -1,50 +1,47 @@
 import Vue from 'vue'
-import VueRouter from 'vue-router'
-
-Vue.use(VueRouter)
+import { createRouter, createWebHistory } from 'vue-router'
 
 const routes = [
   {
     path: '/',
     name: 'NoSidPage',
-    component: () => import(/* webpackChunkName: "NoSidPage" */ '../views/NoSidPage.vue')
+    component: () => import('../views/NoSidPage.vue')
   },
   {
     path: '/select_tracks/sid/:sid',
     alias: '/:sid',
     name: 'SelectTracksPage',
-    component: () => import(/* webpackChunkName: "SelectTracksPage" */ '../views/SelectTracksPage.vue'),
+    component: () => import('../views/SelectTracksPage.vue'),
     props: true
   },
   {
     path: '/toverview/sid/:sid',
     name: 'TrackOverviewPage',
-    component: () => import(/* webpackChunkName: "TrackOverViewPage" */ '../views/TrackOverviewPage.vue'),
+    component: () => import('../views/TrackOverviewPage.vue'),
     props: true
   },
   {
     path: '/upload/sid/:sid',
     name: 'UploadPage',
-    component: () => import(/* webpackChunkName: "UploadPage" */ '../views/UploadPage.vue'),
+    component: () => import('../views/UploadPage.vue'),
     props: true
   },
   {
     path: '/track_multi_edit/sid/:sid',
     name: 'TrackMultiEditPage',
-    component: () => import(/* webpackChunkName: "TrackMultiEdit" */ '../views/TrackMultiEditPage.vue'),
+    component: () => import('../views/TrackMultiEditPage.vue'),
     props: true
   },
   {
     path: '/track/:id/sid/:sid',
     name: 'TrackDetailPage',
-    component: () => import(/* webpackChunkName: "TrackDetail" */ '../views/TrackDetailPage.vue'),
+    component: () => import('../views/TrackDetailPage.vue'),
     props: true
   }
 ]
 
-const router = new VueRouter({
-  mode: 'history',
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(import.meta.env.BASE_URL),
   routes
 })
 
